pyproject.toml
[project]
name = "langchain"
version = "0.1.0"
description = "chatbot with langchain openai"
requires-python = ">=3.11"
dependencies = [
    "duckduckgo-search>=7.4.5",
    "langchain>=0.3.19",
    "langchain-community>=0.3.18",
    "langchain-openai>=0.3.6",
    "python-dotenv>=1.0.1",
]
